requirements.txt
# Core RAG
openai>=1.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# Data handling
requests>=2.31.0
lxml>=5.2.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0

# Hallucination detection
rouge-score>=0.1.2
nltk>=3.8.1
transformers>=4.40.0
torch>=2.2.0

# Config & environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.7.0

# Experiment tracking
wandb>=0.17.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
mutmut>=2.4.0

# Utilities
colorlog>=6.8.0
rich>=13.7.0
